+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Version 0.2.1 (2019-07-18)
 - FEAT: implement cloudwatch get-metric-data
 - ENH: move cw_getMetricStatistics out of ec2op to here for later integration (check issue #4)
 - ENH: use package name "git-remote-aws" with dashes instead of "gitRemoteAws" in setup.py
+- ENH: add locale at top of cli.py files
 
 
 Version 0.2.0 (2019-07-16)